fix deps

+++ geotrek/api/management/commands/sync_mobile.py
@@ -493,7 +493,8 @@ class Command(BaseCommand):
         if os.path.exists(self.dst_root):
             tmp_root2 = os.path.join(os.path.dirname(self.dst_root), 'deprecated_sync_mobile')
             os.rename(self.dst_root, tmp_root2)
-            shutil.rmtree(tmp_root2)
+            if os.path.exists(tmp_root2):
+                shutil.rmtree(tmp_root2)
         os.rename(self.tmp_root, self.dst_root)
         os.chmod(self.dst_root, stat.S_IRWXU | stat.S_IRWXG | stat.S_IROTH | stat.S_IXOTH)
         os.mkdir(self.tmp_root)  # Recreate otherwise python3.6 will complain it does not find the tmp dir at cleanup.
